requirements.txt
spacy
pdfminer.six
pytest